+++ Cargo.toml
@@ -79,4 +79,5 @@ halo2_solidity_verifier = { path = "crates/halo2-verifier" }
 powers-of-tau = { path = "crates/powers-of-tau" }
 shielder-relayer = { path = "crates/shielder-relayer" }
 shielder-rust-sdk = { path = "crates/shielder-rust-sdk" }
+shielder-setup = { path = "crates/shielder-setup" }
 type-conversions = { path = "crates/type-conversions" }
