+++ crates/halo2-verifier/Cargo.toml
@@ -21,5 +21,5 @@ itertools = { workspace = true }
 powers-of-tau = { workspace = true }
 ruint = { workspace = true }
 shielder-circuits = { workspace = true }
-shielder-rust-sdk = { workspace = true, features = ["parameter_generation"] }
+shielder-setup = { workspace = true }
 type-conversions = { workspace = true }
