+++ crates/shielder-relayer/src/metrics.rs
@@ -7,8 +7,8 @@ use axum::{
     response::IntoResponse,
 };
 use metrics_exporter_prometheus::{Matcher, PrometheusBuilder, PrometheusHandle};
-use shielder_rust_sdk::{alloy_primitives::U256, native_token::NATIVE_TOKEN_DECIMALS};
-
+use shielder_rust_sdk::{alloy_primitives::U256};
+use shielder_setup::native_token::NATIVE_TOKEN_DECIMALS;
 use crate::monitor::{healthy, Balances};
 
 pub const TOTAL_REQUESTS_METRIC: &str = "http_requests_total";
