+++ crates/shielder-relayer/src/relay/mod.rs
@@ -8,10 +8,9 @@ use shielder_relayer::{server_error, RelayQuery, RelayResponse, SimpleServiceRes
 use shielder_rust_sdk::{
     alloy_primitives::{Address, U256},
     contract::ShielderContract::withdrawNativeCall,
-    version::{contract_version, ContractVersion},
 };
 use tracing::{debug, error};
-
+use shielder_setup::version::{contract_version, ContractVersion};
 pub use crate::relay::taskmaster::Taskmaster;
 use crate::{
     metrics::WITHDRAW_FAILURE,
