+++ crates/shielder-relayer/src/relay/request_trace.rs
@@ -4,10 +4,9 @@ use shielder_relayer::RelayQuery;
 use shielder_rust_sdk::{
     alloy_primitives::{Address, TxHash, U256},
     contract::ShielderContractError,
-    version::ContractVersion,
 };
 use tracing::{error, info};
-
+use shielder_setup::version::ContractVersion;
 use crate::metrics::{WITHDRAW_DRY_RUN_FAILURE, WITHDRAW_FAILURE, WITHDRAW_SUCCESS};
 
 type Measurement = (String, Duration);
