+++ crates/shielder-rust-sdk/Cargo.toml
@@ -25,6 +25,7 @@ rand = { workspace = true, features = ["small_rng"], optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 sha3 = { workspace = true, optional = true }
 shielder-circuits = { workspace = true, optional = true }
+shielder-setup = { workspace = true }
 static_assertions = { workspace = true, optional = true }
 thiserror = { workspace = true, optional = true }
 tracing = { workspace = true, optional = true }
@@ -60,8 +61,3 @@ contract = [
     "tracing",
     "type-conversions",
 ]
-native_token = []
-parameter_generation = [
-    "rand"
-]
-eip_155 = []
