+++ crates/shielder-rust-sdk/src/account/call_data.rs
@@ -11,6 +11,7 @@ use shielder_circuits::{
     withdraw::WithdrawProverKnowledge,
     Field, ProverKnowledge, PublicInputProvider, F,
 };
+use shielder_setup::version::{contract_version, ContractVersion};
 use type_conversions::{field_to_u256, u256_to_field};
 
 use super::secrets::id_hiding_nonce;
@@ -20,7 +21,6 @@ use crate::{
         ShielderContract::{depositNativeCall, newAccountNativeCall, withdrawNativeCall},
         WithdrawCommitment,
     },
-    version::{contract_version, ContractVersion},
 };
 
 struct ActionSecrets {
