+++ crates/shielder-rust-sdk/src/contract/merkle_path.rs
@@ -1,7 +1,6 @@
 use alloy_primitives::U256;
-
+use shielder_setup::consts::{ARITY, TREE_HEIGHT};
 use crate::{
-    consts::{ARITY, TREE_HEIGHT},
     contract::{call_type::DryRun, ContractResult, ShielderUser},
 };
