+++ crates/shielder-setup/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "shielder-setup"
+version = "0.1.0"
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+license.workspace = true
+categories.workspace = true
+repository.workspace = true
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+alloy-primitives = { workspace = true }
+rand = { workspace = true }
+shielder-circuits = { workspace = true }
