+++ crates/test-ts-conversions/Cargo.toml
@@ -12,5 +12,6 @@ repository.workspace = true
 clap = { workspace = true, features = ["derive"] }
 halo2_proofs = { workspace = true }
 shielder-circuits = { workspace = true }
-shielder-rust-sdk = { workspace = true, features = ["account", "parameter_generation"] }
+shielder-rust-sdk = { workspace = true, features = ["account"] }
+shielder-setup = { workspace = true }
 type-conversions = { workspace = true }
